Add services
I refactored data access to the HeroService class.
I registered the HeroService as the provider of its service at the root level so that it can be injected anywhere in the app.
I used Angular Dependency Injection to inject it into a component.
I gave the HeroService get data method an asynchronous signature.
I discovered Observable and the RxJS Observable library.
I used RxJS of() to return an observable of mock heroes (Observable<Hero[]>).
The component's ngOnInit lifecycle hook calls the HeroService method, not the constructor.
I created a MessageService for loosely-coupled communication between classes.
The HeroService injected into a component is created with another injected service, MessageService.